Sistema funcional con interrupciones

+++ Código/Control_Temperatura/Control_Temperatura/termometro.h
@@ -6,5 +6,5 @@
  */ 
 #include <avr/io.h>
 void TERMOMETRO_init(void);
-uint8_t TERMOMETRO_get_temperatura_entero(void);
+int TERMOMETRO_get_temperatura_entero(void);
 float TERMOMETRO_get_temperatura_real(void);
